# api/openapi.yaml
openapi: 3.1.0
info:
  description: REST and Server-Sent Events API of the Sift server. Errors are RFC
    9457 problem details (`application/problem+json`).
  title: Sift Server API
  version: v1
security:
- bearerAuth: []
...
paths:
...
  /api/v1/repositories:
    get:
      operationId: repositoryList
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepositoryResponse"
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
      summary: List repositories
      tags:
      - repositories
    post:
      operationId: repositoryCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepositoryRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryResponse"
          description: Repository created; `Location` points at it
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation failed
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: A repository with the same name exists
      summary: Register a repository
      tags:
      - repositories
  /api/v1/repositories/{id}:
    delete:
      operationId: repositoryDelete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Repository deleted
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Repository not found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Repository still has active agent runs
      summary: Delete a repository
      tags:
      - repositories
    get:
      operationId: repositoryGet
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryResponse"
          description: The repository
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Repository not found
      summary: Get a repository
      tags:
      - repositories
    put:
      description: Replaces the URL and/or rotates or clears the token.
      operationId: repositoryUpdate
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepositoryRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryResponse"
          description: The updated repository
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation failed
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Repository not found
      summary: Update a repository
      tags:
      - repositories
  /api/v1/results:
    get:
      operationId: reviewResultList
      parameters:
      - in: query
        name: repositoryUrl
        required: false
        schema:
          type: string
      - in: query
        name: commitSha
        required: false
        schema:
          type: string
      - in: query
        name: agentRunId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseReviewResultSummaryResponse"
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
      summary: List review results
      tags:
      - results
  /api/v1/results/{id}:
    get:
      operationId: reviewResultGet
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResultResponse"
          description: The result including its findings
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Result not found
      summary: Get a review result with all its findings
      tags:
      - results
  /api/v1/results/{id}/findings:
    get:
      description: Optionally narrowed by severity and file
      operationId: reviewResultFindings
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: severity
        required: false
        schema:
          type: string
          enum:
          - BLOCKER
          - MAJOR
          - MINOR
          - INFO
      - in: query
        name: file
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewFindingResponse"
          description: Matching findings
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Result not found
      summary: List the findings of a review result
      tags:
      - results
components:
  schemas:
...
    CreateRepositoryRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        token:
          type:
          - string
          - "null"
        url:
          type: string
          minLength: 1
      required:
      - name
      - url
...
    PageResponseReviewResultSummaryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReviewResultSummaryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
    RepositoryResponse:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        hasToken:
          type: boolean
        id:
          type: string
          format: uuid
        name:
          type: string
        secretName:
          type:
          - string
          - "null"
        updatedAt:
          type: string
          format: date-time
        url:
          type: string
    ReviewFindingResponse:
      type: object
      properties:
        category:
          type:
          - string
          - "null"
        endLine:
          type:
          - integer
          - "null"
          format: int32
        file:
          type: string
        id:
          type: integer
          format: int64
        message:
          type: string
        severity:
          type: string
          enum:
          - BLOCKER
          - MAJOR
          - MINOR
          - INFO
        startLine:
          type:
          - integer
          - "null"
          format: int32
        suggestion:
          type:
          - string
          - "null"
    ReviewResultResponse:
      type: object
      properties:
        agentRunId:
          type:
          - string
          - "null"
          format: uuid
        baseBranch:
          type: string
        branch:
          type: string
        commitSha:
          type: string
        completedAt:
          type: string
          format: date-time
        executionId:
          type: string
        findingCount:
          type: integer
          format: int64
        findings:
          type: array
          items:
            $ref: "#/components/schemas/ReviewFindingResponse"
        id:
          type: string
          format: uuid
        pullRequest:
          type:
          - string
          - "null"
        receivedAt:
          type: string
          format: date-time
        repositoryUrl:
          type: string
        summary:
          type: string
    ReviewResultSummaryResponse:
      type: object
      properties:
        agentRunId:
          type:
          - string
          - "null"
          format: uuid
        baseBranch:
          type: string
        branch:
          type: string
        commitSha:
          type: string
        completedAt:
          type: string
          format: date-time
        executionId:
          type: string
        findingCount:
          type: integer
          format: int64
        id:
          type: string
          format: uuid
        pullRequest:
          type:
          - string
          - "null"
        receivedAt:
          type: string
          format: date-time
        repositoryUrl:
          type: string
        summary:
          type: string
...
    UpdateRepositoryRequest:
      type: object
      properties:
        clearToken:
          type: boolean
        token:
          type:
          - string
          - "null"
        url:
          type:
          - string
          - "null"
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Access token issued by the OIDC provider announced by `GET /api/v1/auth/config`.
      scheme: bearer
      type: http